+++ plugin/src/main/kotlin/gradle/greeting/GradleGreetingPlugin.kt
@@ -3,19 +3,17 @@
  */
 package gradle.greeting
 
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-
-/**
- * A simple 'hello world' plugin.
- */
-class GradleGreetingPlugin : Plugin<Project> {
-    override fun apply(project: Project) {
-        // Register a task
-        project.tasks.register("greeting") { task ->
-            task.doLast {
-                println("Hello from plugin 'gradle.greeting.greeting'")
-            }
-        }
-    }
-}
+//
+///**
+// * A simple 'hello world' plugin.
+// */
+//class GradleGreetingPlugin : Plugin<Project> {
+//    override fun apply(project: Project) {
+//        // Register a task
+//        project.tasks.register("greeting") { task ->
+//            task.doLast {
+//                println("Hello from plugin 'gradle.greeting.greeting'")
+//            }
+//        }
+//    }
+//}
